+++ imports/ui/administration/AdministrationPage.vue
@@ -3,11 +3,18 @@
     <template v-if="$subReady.user">
       <template v-if="isAdmin()">
         <v-tabs>
+          <v-tabs-slider color="accent" />
           <v-tab id="tab-users">
-            Utilisateurs
+            {{ $t('Users') }}
+          </v-tab>
+          <v-tab id="tab-users">
+            {{ $t('BPMN examples') }}
           </v-tab>
           <v-tab-item>
             <administration-users />
+          </v-tab-item>
+          <v-tab-item>
+            <administration-bpmn-examples />
           </v-tab-item>
         </v-tabs>
       </template>
@@ -18,9 +25,14 @@
 <script>
 import { Permissions } from "/imports/api/permissions/permissions";
 
-export default {
-  props: {},
+import AdministrationUsers from "./users/AdministrationUsers";
+import AdministrationBpmnExamples from "./bpmn/AdministrationBpmnExamples";
 
+export default {
+  components: {
+    AdministrationUsers,
+    AdministrationBpmnExamples
+  },
   data() {
     return {};
   },