+++ imports/ui/administration/bpmn/AdministrationBpmnExamples.vue
@@ -1,0 +1,16 @@
+<template>
+  <div>
+    <examples-list />
+  </div>
+</template>
+
+<script>
+import ExamplesList from "/imports/ui/bpmn/examples/ExamplesList";
+
+export default {
+  name: "AdministrationBpmnExamples",
+  components: {
+    ExamplesList
+  }
+};
+</script>